PowerTrimmingHistograms update, time is in HH:MM:SS format

See merge request cmsinnertracker/Ph2_ACF!377

+++ DQMUtils/RD53PowerTrimmingHistograms.h
@@ -19,7 +19,7 @@
 
 struct PowerTrimmingData
 {
-    float    timestamp;
+    double    timestamp;
     uint16_t bit;
     float    ChipCurrent;
     float    ANA_IN_CURR;
